merge

+++ Sources/DeepLearning/Operators.swift
@@ -35,7 +35,7 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
     // TODO: Verify that these calculations are correct.
     @inlinable
     internal func _vjpBatchNormalized(
-        alongAxis axis: Int32,
+        alongAxis axis: Int,
         offset: Tensor,
         scale: Tensor,
         epsilon: Scalar
@@ -84,7 +84,7 @@ public extension Tensor where Scalar: BinaryFloatingPoint {
         where Scalar : TensorFlowFloatingPoint
     )
     func batchNormalized(
-        alongAxis axis: Int32,
+        alongAxis axis: Int,
         offset: Tensor = Tensor(0),
         scale: Tensor = Tensor(1),
         epsilon: Scalar = 0.001
@@ -135,14 +135,14 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
     internal func conv2DBackpropInput(
         shape: Tensor<Int32>,
         filter: Tensor,
-        strides: (Int32, Int32, Int32, Int32),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> Tensor {
         return Raw.conv2DBackpropInput(
             inputSizes: shape,
             filter: filter,
             outBackprop: self,
-            strides: [strides.0, strides.1, strides.2, strides.3],
+            strides: [Int32(strides.0), Int32(strides.1), Int32(strides.2), Int32(strides.3)],
             padding: padding.raw2,
             explicitPaddings: [])
     }
@@ -153,14 +153,14 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
     internal func conv2DBackpropFilter(
         input: Tensor,
         filterSizes: Tensor<Int32>,
-        strides: (Int32, Int32, Int32, Int32),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> Tensor {
         return Raw.conv2DBackpropFilter(
             input,
             filterSizes: filterSizes,
             outBackprop: self,
-            strides: [strides.0, strides.1, strides.2, strides.3],
+            strides: [Int32(strides.0), Int32(strides.1), Int32(strides.2), Int32(strides.3)],
             padding: padding.raw2,
             explicitPaddings: [])
     }
@@ -169,7 +169,7 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
     internal func _vjpConv2DBackpropInput(
         _ shape: Tensor<Int32>,
         _ filter: Tensor,
-        _ strides: (Int32, Int32, Int32, Int32),
+        _ strides: (Int, Int, Int, Int),
         _ padding: Padding
     ) -> (Tensor, (Tensor) -> (Tensor, Tensor)) {
         let value = conv2DBackpropInput(shape: shape, filter: filter, strides: strides,
@@ -187,7 +187,7 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
     internal func _vjpConv2DBackpropFilter(
         _ input: Tensor,
         _ filterSizes: Tensor<Int32>,
-        _ strides: (Int32, Int32, Int32, Int32),
+        _ strides: (Int, Int, Int, Int),
         _ padding: Padding
     ) -> (Tensor, (Tensor) -> (Tensor, Tensor)) {
         let value = conv2DBackpropFilter(input: input, filterSizes: filterSizes,
@@ -204,7 +204,7 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
     @inlinable
     internal func _vjpConvolved2D(
         filter: Tensor,
-        strides: (Int32, Int32, Int32, Int32),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> (Tensor, (Tensor) -> (Tensor, Tensor)) {
         let value = convolved2D(withFilter: filter, strides: strides,
@@ -225,8 +225,8 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
 
     @inlinable
     internal func _vjpMaxPooled(
-        kernelSize: (Int32, Int32, Int32, Int32),
-        strides: (Int32, Int32, Int32, Int32),
+        kernelSize: (Int, Int, Int, Int),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> (Tensor, (Tensor) -> Tensor) {
         // TODO: Currently this is not higher order differentiable. Redefine in
@@ -238,8 +238,10 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
                 origInput: self,
                 origOutput: value,
                 grad: v,
-                ksize: Tensor<Int32>([kernelSize.0, kernelSize.1, kernelSize.2, kernelSize.3]),
-                strides: Tensor<Int32>([strides.0, strides.1, strides.2, strides.3]),
+                ksize: Tensor<Int32>([Int32(kernelSize.0), Int32(kernelSize.1),
+                                      Int32(kernelSize.2), Int32(kernelSize.3)]),
+                strides: Tensor<Int32>([Int32(strides.0), Int32(strides.1),
+                                        Int32(strides.2), Int32(strides.3)]),
                 padding: padding.raw
             )
         })
@@ -247,8 +249,8 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
 
     @inlinable
     internal func _vjpAveragePooled(
-        kernelSize: (Int32, Int32, Int32, Int32),
-        strides: (Int32, Int32, Int32, Int32),
+        kernelSize: (Int, Int, Int, Int),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> (Tensor, (Tensor) -> Tensor) {
         // TODO: Currently this is not higher order differentiable. Redefine in
@@ -259,8 +261,9 @@ public extension Tensor where Scalar: TensorFlowFloatingPoint {
             return Raw.avgPoolGrad(
                 origInputShape: self.shapeTensor,
                 grad: v,
-                ksize: [kernelSize.0, kernelSize.1, kernelSize.2, kernelSize.3],
-                strides: [strides.0, strides.1, strides.2, strides.3],
+                ksize: [Int32(kernelSize.0), Int32(kernelSize.1),
+                        Int32(kernelSize.2), Int32(kernelSize.3)],
+                strides: [Int32(strides.0), Int32(strides.1), Int32(strides.2), Int32(strides.3)],
                 padding: padding.raw
             )
         })
@@ -285,13 +288,13 @@ public extension Tensor where Scalar: FloatingPoint {
     )
     func convolved2D(
         withFilter filter: Tensor,
-        strides: (Int32, Int32, Int32, Int32),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> Tensor {
         return Raw.conv2D(
             self,
             filter: filter,
-            strides: [strides.0, strides.1, strides.2, strides.3],
+            strides: [Int32(strides.0), Int32(strides.1), Int32(strides.2), Int32(strides.3)],
             padding: padding.raw2,
             explicitPaddings: [])
     }
@@ -310,16 +313,16 @@ public extension Tensor where Scalar: FloatingPoint {
         where Scalar : TensorFlowFloatingPoint
     )
     func maxPooled(
-        kernelSize: (Int32, Int32, Int32, Int32),
-        strides: (Int32, Int32, Int32, Int32),
+        kernelSize: (Int, Int, Int, Int),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> Tensor {
         return Raw.maxPoolV2(
             self,
-            ksize: Tensor<Int32>([kernelSize.0, kernelSize.1,
-                                 kernelSize.2, kernelSize.3]),
-            strides: Tensor<Int32>([strides.0, strides.1,
-                                   strides.2, strides.3]),
+            ksize: Tensor<Int32>([Int32(kernelSize.0), Int32(kernelSize.1),
+                                  Int32(kernelSize.2), Int32(kernelSize.3)]),
+            strides: Tensor<Int32>([Int32(strides.0), Int32(strides.1),
+                                    Int32(strides.2), Int32(strides.3)]),
             padding: padding.raw)
     }
 
@@ -337,14 +340,15 @@ public extension Tensor where Scalar: FloatingPoint {
         where Scalar : TensorFlowFloatingPoint
     )
     func averagePooled(
-        kernelSize: (Int32, Int32, Int32, Int32),
-        strides: (Int32, Int32, Int32, Int32),
+        kernelSize: (Int, Int, Int, Int),
+        strides: (Int, Int, Int, Int),
         padding: Padding
     ) -> Tensor {
         return Raw.avgPool(
             value: self,
-            ksize: [kernelSize.0, kernelSize.1, kernelSize.2, kernelSize.3],
-            strides: [strides.0, strides.1, strides.2, strides.3],
+            ksize: [Int32(kernelSize.0), Int32(kernelSize.1),
+                    Int32(kernelSize.2), Int32(kernelSize.3)],
+            strides: [Int32(strides.0), Int32(strides.1), Int32(strides.2), Int32(strides.3)],
             padding: padding.raw)
     }
 }
